move proposed dimensions into primitives, add From impl for embedded-graphics Size
