+++ src/view/modifier/priority.rs
@@ -1,7 +1,7 @@
 use crate::{
     environment::{LayoutEnvironment, RenderEnvironment},
-    layout::{Layout, ProposedDimensions, ResolvedLayout},
-    primitives::Point,
+    layout::{Layout, ResolvedLayout},
+    primitives::{Point, ProposedDimensions},
     render::CharacterRender,
     render_target::CharacterRenderTarget,
 };
