+++ src/view/shape/rounded_rectangle.rs
@@ -1,6 +1,6 @@
 use crate::{
-    layout::{Layout, ProposedDimensions, ResolvedLayout},
-    primitives::Point,
+    layout::{Layout, ResolvedLayout},
+    primitives::{Point, ProposedDimensions},
 };
 
 #[derive(Debug, Copy, Clone, PartialEq, Default)]
